% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barycentric.R
\name{calc_barycentric_coords}
\alias{calc_barycentric_coords}
\title{Title}
\usage{
calc_barycentric_coords(x, y, x1, y1, x2, y2, x3, y3)
}
\arguments{
\item{x}{}

\item{y}{}

\item{x1}{}

\item{y1}{}

\item{x2}{}

\item{y2}{}

\item{x3}{}

\item{y3}{}
}
\value{

}
\description{
Title
}
