% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_conductance.R
\name{calc_conductance_MHEq3}
\alias{calc_conductance_MHEq3}
\title{Calculate conductance using Mehl & Hill (2010) Equation 3}
\usage{
calc_conductance_MHEq3(
  swdf,
  k_str,
  kv_node,
  str_width,
  str_thick,
  node_thick,
  node_area
)
}
\arguments{
\item{swdf}{DataFrame/Geometry of node barycentric coordinates as returned by
\code{\link{calc_stream_voronoi_weights}}}

\item{k_str}{numeric, streambed hydraulic conductivity}

\item{kv_node}{numeric, vertical hydraulic conductivity of the aquifer material}

\item{str_width}{numeric, stream width}

\item{str_thick}{numeric, streambed thickness (head loss zone)}

\item{node_thick}{numeric, thickness of finite difference cell (node)}

\item{node_area}{Planar cross-sectional area of cell (node)}
}
\value{
numeric, conductance
}
\description{
Vectorized, non-constant values should be same length as swdf.
}
\author{
Leland Scantlebury
}
