% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_conductance.R
\name{calc_conductance_modflow}
\alias{calc_conductance_modflow}
\title{Title}
\usage{
calc_conductance_modflow(swdf, k_streambed, str_width, thickness)
}
\arguments{
\item{stream_thick}{}
}
\value{

}
\description{
Title
}
