% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_conductance.R
\name{calc_conductance_modflow_perLen}
\alias{calc_conductance_modflow_perLen}
\title{Calculate classic MODFLOW-style stream conductance (per unit length)
Per Unit Length Conductance = width x streambed hydraulic conductivity / streambed thickness}
\usage{
calc_conductance_modflow_perLen(k_str, str_width, str_thick)
}
\arguments{
\item{k_str}{numeric, streambed hydraulic conductivity}

\item{str_width}{numeric, stream width}

\item{str_thick}{numeric, streambed thickness (head loss zone)}
}
\value{
numeric, conductance
}
\description{
Calculate classic MODFLOW-style stream conductance (per unit length)
Per Unit Length Conductance = width x streambed hydraulic conductivity / streambed thickness
}
\examples{
#-- Read in shapefiles
str <- read_sf(system.file("extdata", "MehlandHill2010_stream.shp", package = "MFUSGPBJ"))
tri <- read_sf(system.file("extdata", "720_triangles.shp", package = "MFUSGPBJ"))
vor <- read_sf(system.file("extdata", "720_voronoi.shp", package = "MFUSGPBJ"))
str <- line_explode(str)

#-- Calculate barycentric weight DF
swdf <- calc_stream_voronoi_weights(stream = str, voronoi = vor, triangles = tri)

#-- Calculate distances
swdf <- stream_elev_from_slope(swdf = swdf, slope = 0.0015, initial_elev = 50)

#-- Calculate conductances
swdf$Conductance <- calc_conductance_modflow_perLen(k_str = 1, str_width = 1,
                                                    thickness = 0.5)
}
\author{
Leland Scantlebury
}
