% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_drains_by_length.R
\name{clean_drains}
\alias{clean_drains}
\title{Clean Drain Segments by Length & Combine Co-located Segments}
\usage{
clean_drains(drndf, min_seg_length = 1e-04, verbose = T)
}
\arguments{
\item{drndf}{Output from \code{\link{calc_stream_drn}}. Must be ordered and contain geometry.}

\item{min_seg_length}{Minimum allowed segment length (default: 1e-4)}

\item{verbose}{T/F if information should be relayed through messages (default: True).}
}
\value{

}
\description{
Clean Drain Segments by Length & Combine Co-located Segments
}
\examples{
#-- Read in shapefiles
str <- read_sf(system.file("extdata", "straight_river.shp", package = "MFUSGPBJ"))
vor <- read_sf(system.file("extdata", "straight_voronoi.shp", package = "MFUSGPBJ"))

#-- Explode polyline
str <- line_explode(str)

#-- Create drndf
drndf <- calc_stream_drn(stream = stream, voronoi = vor)

#-- Clean up
drndf <- clean_drains(drndf, min_seg_length=0.5)
}
