% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stream_elev_from_raster.R
\name{extract_stream_elev_from_raster}
\alias{extract_stream_elev_from_raster}
\title{Extract Stream Elevations from Raster (DEM)}
\usage{
extract_stream_elev_from_raster(swdf, dem, method = "simple")
}
\arguments{
\item{swdf}{DataFrame/Geometry of node barycentric coordinates as returned by
\code{\link{calc_stream_voronoi_weights}}}

\item{dem}{Raster object}

\item{method}{character, 'simple' or 'bilinear', passed to raster::extract()}
}
\value{
swdf DataFrame/Geometry with added seg1.elev and seg2.elev columns
}
\description{
Used to get segment start/end elevations from a raster digital elevation model (DEM)
}
