% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_seg_elevs.R
\name{fix_seg_elevs}
\alias{fix_seg_elevs}
\title{fix_seg_elevs}
\usage{
fix_seg_elevs(
  sdf,
  idcol,
  start_id,
  elev_col1 = "seg1.elev",
  elev_col2 = "seg2.elev",
  prev_segids = NULL,
  verbose = T
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
fix_seg_elevs
}
