% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barycentric.R
\name{geo_to_barycentric_coords}
\alias{geo_to_barycentric_coords}
\title{Title}
\usage{
geo_to_barycentric_coords(seg_geo, tri_geo)
}
\arguments{
\item{seg_geo}{}

\item{tri_geo}{}
}
\value{

}
\description{
Title
}
