% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_utils.R
\name{line_explode}
\alias{line_explode}
\title{Convert single polyline into multipolyline segments, based on vertices
Or "explode" the line, as it is commonly referred to in GIS programs}
\usage{
line_explode(polyline)
}
\arguments{
\item{polyline}{sf polyline object}
}
\value{
sf multipolyline object
}
\description{
Warning: This function is NOT robust. Poorly-ordered shapefiles can easily confused it!
Plot the resulting object to ensure it has connected the lines properly. GIS programs
(e.g., QGIS) have much more robust explode processes that will do much better job.
}
\examples{
#-- Read in shapefiles
str <- read_sf(system.file("extdata", "MehlandHill2010_stream.shp", package = "pbjr"))

#-- Explode polyline
str <- line_explode(str)
}
\author{
Leland Scantlebury
}
