% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.segment_domain.R
\name{plot.segment_domain}
\alias{plot.segment_domain}
\title{Title}
\usage{
\method{plot}{segment_domain}(triangles, voronoi, stream, seg_id, tri_id, trim_seg = T, title = F)
}
\arguments{
\item{title}{}
}
\value{

}
\description{
Title
}
