% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapefile_utils.R
\name{reorder_segments}
\alias{reorder_segments}
\title{Reorder line segments after intersection}
\usage{
reorder_segments(src, parts)
}
\arguments{
\item{src}{original line sf polyline object}

\item{parts}{intersected sf polygon object}
}
\value{
Re-ordered version of parts
}
\description{
Reorder line segments after intersection
}
\details{
Very briefly modificied from Spacedman's solution on StackExchange:
\url{https://gis.stackexchange.com/questions/295806/r-turn-off-automatic-ordering-of-linestrings-when-applying-sfst-intersection}
}
\examples{
#-- Read in shapefiles
str <- read_sf(system.file("extdata", "MehlandHill2010_stream.shp", package = "pbjr"))
tri <- read_sf(system.file("extdata", "720_triangles.shp", package = "pbjr"))

# Intersect
tri_stream <- st_intersection(tri, str)

#-- st_intersection can mess up segment order - it uses the triangle ID # to determine the order
tri_stream <- reorder_segments(str, tri_stream)
}
