% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_head_from_slope.R
\name{spec_head_from_slope}
\alias{spec_head_from_slope}
\title{Calculate specified head elevations given geometry & slope
Assumes a single line.}
\usage{
spec_head_from_slope(swdf, sp, slope, initial_elev, order_by_index = F)
}
\arguments{
\item{swdf}{DataFrame/Geometry of node barycentric coordinates as returned by
\code{\link{calc_stream_voronoi_weights}}}

\item{sp}{integer, stress period of heads, used for column naming}

\item{slope}{numeric, slope}

\item{initial_elev}{streambed elevation at start of segment 1}
}
\value{
swdf DataFrame/Geometry with added seg1.head[sp] and seg1.head[sp] columns
}
\description{
Calculate specified head elevations given geometry & slope
Assumes a single line.
}
\examples{
#-- Read in shapefiles
str <- read_sf(system.file("extdata", "MehlandHill2010_stream.shp", package = "pbjr"))
tri <- read_sf(system.file("extdata", "720_triangles.shp", package = "pbjr"))
vor <- read_sf(system.file("extdata", "720_voronoi.shp", package = "pbjr"))
str <- line_explode(str)

#-- Calculate barycentric weight DF
swdf <- calc_stream_voronoi_weights(stream = str, voronoi = vor, triangles = tri)

#-- Calculate distances
swdf <- spec_head_from_slope(swdf = swdf, sp = 1, slope = 0.0015, initial_elev = 50)
}
\seealso{
\code{\link{stream_elev_from_slope}}
}
\author{
Leland Scantlebury
}
