% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_elev_from_slope.R
\name{stream_elev_from_slope}
\alias{stream_elev_from_slope}
\title{Title}
\usage{
stream_elev_from_slope(swdf, slope, initial_elev)
}
\arguments{
\item{initial_elev}{}
}
\value{

}
\description{
Title
}
