% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_corners.R
\name{triangle_corners}
\alias{triangle_corners}
\title{Triangle polygon to triangle point sf dataframe}
\usage{
triangle_corners(triangles)
}
\arguments{
\item{triangles}{sf polygon object of triangle(s)}
}
\value{
dataframe of three sf point geometry objects, with ID column for reference
}
\description{
Triangle polygon to triangle point sf dataframe
}
\examples{
# Read in shapefile
tri <- read_sf(system.file("extdata", "720_triangles.shp", package = "MFUSGPBJ"))

# Convert to corner coordinates
tricorner <- triangle_corners(tri)
}
\author{
Leland Scantlebury
}
