% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.PBJpackage.R
\name{write.PBJpackage}
\alias{write.PBJpackage}
\title{Title}
\usage{
write.PBJpackage(swdf, filename, head_interp = "GEO", seg_sort = T)
}
\arguments{
\item{seg_sort}{}
}
\value{

}
\description{
Title
}
