% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.PBJpackage_single.R
\name{write.PBJpackage_single}
\alias{write.PBJpackage_single}
\title{Title}
\usage{
write.PBJpackage_single(
  segnodedf,
  filename,
  nSPs,
  IPBJCB,
  pbjmode = "DRAIN",
  condtype = "UNITCOND",
  NodeCol = "node",
  SPwarnings = T
)
}
\arguments{
\item{SPwarnings}{}
}
\value{

}
\description{
Title
}
