% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barycentric.R
\name{calc_barycentric_line_coords}
\alias{calc_barycentric_line_coords}
\title{Calculate barycentric coordinates at given line start/end points
Vectorized function to get barycentric coodinates for segment start/end points for all triangles}
\usage{
calc_barycentric_line_coords(lx1, ly1, lx2, ly2, x1, y1, x2, y2, x3, y3)
}
\arguments{
\item{lx1}{numeric (or array), line first x-coordinate}

\item{ly1}{numeric (or array), line first y-coordinate}

\item{lx2}{numeric (or array), line second x-coordinate}

\item{ly2}{numeric (or array), line second y-coordinate}

\item{x1}{numeric (or array), triangle corner one x-coordinate}

\item{y1}{numeric (or array), triangle corner one y-coordinate}

\item{x2}{numeric (or array), triangle corner two x-coordinate}

\item{y2}{numeric (or array), triangle corner two y-coordinate}

\item{x3}{numeric (or array), triangle corner three x-coordinate}

\item{y3}{numeric (or array), triangle corner three y-coordinate}
}
\value{
matrix of barycentric coordinates for seg1 & seg2 (start & end)
}
\description{
Calculate barycentric coordinates at given line start/end points
Vectorized function to get barycentric coodinates for segment start/end points for all triangles
}
\examples{
calc_barycentric_line_coords(lx1 = c(523300,523400),
                             ly1 = c(4925787.45,4925802.08),
                             lx2 = c(523305,523405),
                             ly2 = c(4925790.45,4925888.08),
                             x1 = 523232.75, y1 = 4925893,
                             x2 = 523448.87, y2 = 4925906.09,
                             x3 = 523353.60, y3 = 4925703.32)
}
\author{
Leland Scantlebury
}
