% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.PBJpackage.R
\name{write.PBJpackage}
\alias{write.PBJpackage}
\title{Writes Modflow-USG Polyline Boundary Junction Package}
\usage{
write.PBJpackage(swdf, filename, head_interp = "GEO", seg_sort = T)
}
\arguments{
\item{swdf}{DataFrame/Geometry of node barycentric coordinates as returned by
\code{\link{calc_stream_voronoi_weights}} with additional 'Conductance' and elevation
('seg1.elev' and 'seg2.elev') columns required}

\item{filename}{character, name/location of output file}

\item{seg_sort}{T/F (optional) whether swdf should be sorted by segment prior to output (default: True)}
}
\description{
Writes Modflow-USG Polyline Boundary Junction Package
}
\examples{
write.PBJpackage
#-- Read in shapefiles
str <- read_sf(system.file("extdata", "MehlandHill2010_stream.shp", package = "MFUSGPBJ"))
tri <- read_sf(system.file("extdata", "720_triangles.shp", package = "MFUSGPBJ"))
vor <- read_sf(system.file("extdata", "720_voronoi.shp", package = "MFUSGPBJ"))
str <- line_explode(str)

#-- Calculate barycentric weight DF
swdf <- calc_stream_voronoi_weights(stream = str, voronoi = vor, triangles = tri)

#-- Calculate distances
swdf <- stream_elev_from_slope(swdf = swdf, slope = 0.0015, initial_elev = 50)

#-- Calculate conductances
swdf$Conductance <- calc_conductance_modflow(swdf, k_streambed = 1,
                                             str_width = 1, thickness = 0.5)
#-- Write package file
write.PBJpackage(swdf, filename = paste0(tempdir(),'/model720.pbj'))
}
